SQL utils refactoring
