+++ plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/model/sql/parser/SQLWordDetector.java
@@ -1,0 +1,38 @@
+/*
+ * DBeaver - Universal Database Manager
+ * Copyright (C) 2010-2017 Serge Rider (serge@jkiss.org)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jkiss.dbeaver.model.sql.parser;
+
+import org.eclipse.jface.text.rules.IWordDetector;
+
+/**
+ * Determines whether a given character is valid as part of an SQL keyword in
+ * the current context.
+ */
+public class SQLWordDetector implements IWordDetector
+{
+
+    @Override
+    public boolean isWordStart(char c) {
+        return Character.isUnicodeIdentifierStart(c);
+    }
+
+    @Override
+    public boolean isWordPart(char c) {
+        return Character.isUnicodeIdentifierPart(c) || c == '$';
+    }
+
+}
