+++ plugins/org.jkiss.dbeaver.ui/META-INF/MANIFEST.MF
@@ -11,6 +11,7 @@ Bundle-ActivationPolicy: lazy
 Export-Package: org.jkiss.dbeaver.runtime,
  org.jkiss.dbeaver.ui,
  org.jkiss.dbeaver.ui.controls,
+ org.jkiss.dbeaver.ui.editors,
  org.jkiss.dbeaver.ui.dialogs,
  org.jkiss.dbeaver.ui.dialogs.exec,
  org.jkiss.dbeaver.ui.notifications,
