+++ plugins/org.jkiss.dbeaver.ui/src/org/jkiss/dbeaver/ui/editors/INonPersistentEditorInput.java
